Moves more logic to DOMEditor.
